+++ packages/dso-toolkit/src/components/document-card/document-card.args.ts
@@ -1,10 +1,11 @@
 import { HandlerFunction } from "@storybook/addon-actions";
 import { ArgTypes } from "@storybook/types";
 
-import { noControl } from "../../storybook/index.js";
+import { noControl } from "../../storybook";
 import { DocumentCard } from "./document-card.models.js";
-import { Label } from "../label/label.models.js";
-import { Toggletip } from "../toggletip/toggletip.models.js";
+import { Label } from "../label";
+import { Toggletip } from "../toggletip";
+import { Badge } from "../badge";
 
 export interface DocumentCardArgs {
   label: string;
@@ -13,17 +14,24 @@ export interface DocumentCardArgs {
   typeToelichting: Toggletip<never>;
   meta: Label;
   status: string;
+  statusToelichtingOutline: Badge;
+  statusToelichtingWarning: Badge;
   dsoDocumentCardClick: HandlerFunction;
 }
 
-export const documentCardArgs: Omit<DocumentCardArgs, "meta" | "typeToelichting" | "dsoDocumentCardClick"> = {
+export const documentCardArgs: Omit<
+  DocumentCardArgs,
+  "meta" | "typeToelichting" | "dsoDocumentCardClick" | "statusToelichtingOutline" | "statusToelichtingWarning"
+> = {
   href: "#",
   label: "Omgevingsplan Amsterdam",
   status: "Vastgesteld 07-06-2024 - geheel onherroepelijk in werking",
   active: false,
 };
 
-export const documentCardArgTypes: ArgTypes<Omit<DocumentCardArgs, "meta" | "typeToelichting">> = {
+export const documentCardArgTypes: ArgTypes<
+  Omit<DocumentCardArgs, "meta" | "typeToelichting" | "statusToelichtingOutline" | "statusToelichtingWarning">
+> = {
   label: {
     control: {
       type: "text",
@@ -61,6 +69,8 @@ export function documentCardArgsMapper<TemplateFnReturnType>(
     active: a.active,
     typeItems,
     status: a.status,
+    statusToelichtingOutline: a.statusToelichtingOutline,
+    statusToelichtingWarning: a.statusToelichtingWarning,
     meta: a.meta,
     dsoDocumentCardClick: (e) => a.dsoDocumentCardClick(e.detail),
   };