+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Document Card: Aanvullende functionaliteit ([#2835](https://github.com/dso-toolkit/dso-toolkit/issues/2835))
+
 ## ✍️ Release 67.0.0 - 2024-11-25
 
 ### Removed