+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * Document Card: Aanvullende functionaliteit ([#2835](https://github.com/dso-toolkit/dso-toolkit/issues/2835))
 
+### Changed
+* Header: ShadowRoot ([#2365](https://github.com/dso-toolkit/dso-toolkit/issues/2365))
+
 ## ✍️ Release 67.0.0 - 2024-11-25
 
 ### Removed