+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+## 🔵 Release 67.0.0-ghi-2365.0 - 2024-11-29
+
 ### Added
 * Document Card: Aanvullende functionaliteit ([#2835](https://github.com/dso-toolkit/dso-toolkit/issues/2835))